+++ android/app/src/main/java/com/now/naaga/data/repository/DefaultPlaceRepository.kt
@@ -5,7 +5,7 @@ import com.now.domain.model.Place
 import com.now.domain.repository.PlaceRepository
 import com.now.naaga.data.mapper.toDomain
 import com.now.naaga.data.remote.retrofit.service.PlaceService
-import com.now.naaga.util.getValueOrThrow
+import com.now.naaga.util.extension.getValueOrThrow
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
