+++ android/app/src/main/java/com/now/naaga/util/singleliveevent/SingleLiveData.kt
@@ -1,4 +1,4 @@
-package com.now.naaga.util
+package com.now.naaga.util.singleliveevent
 
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.MutableLiveData
